move helper to correct api, change server list
