+++ src/common/constants/servers.json
@@ -1,1 +1,1 @@
-["https://api.deathwing.me", "https://rpc.ecency.com", "https://api.hive.blog"]
+["https://rpc.ecency.com", "https://api.hive.blog", "https://anyx.io"]
